third_party/WebKit/LayoutTests/fast/css/invalidation/targeted-class-whole-subtree-expected.txt

Check that all children of inner are recalculated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(innerDivs[i], null).backgroundColor is transparent
PASS getComputedStyle(innerDivs[i], null).backgroundColor is transparent
PASS getComputedStyle(innerDivs[i], null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 4
PASS getComputedStyle(innerDivs[i], null).backgroundColor is green
PASS getComputedStyle(innerDivs[i], null).backgroundColor is green
PASS getComputedStyle(innerDivs[i], null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
